+++ ServerProject/src/main/java/com/example/vavaplanit/model/repetition/Repetition.java
@@ -9,7 +9,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Repetition {
-    private Long eventId;
+    private Integer eventId;
     private LocalDate start;
     private LocalDate end;
     private int repetitionInterval;
@@ -17,7 +17,7 @@ public class Repetition {
     public Repetition() {
     }
 
-    public Repetition(Long eventId, LocalDate start, LocalDate end, int repetitionInterval) {
+    public Repetition(Integer eventId, LocalDate start, LocalDate end, int repetitionInterval) {
         this.eventId = eventId;
         this.start = start;
         this.end = end;
@@ -48,11 +48,11 @@ public class Repetition {
         this.repetitionInterval = repetitionInterval;
     }
 
-    public Long getEventId() {
+    public Integer getEventId() {
         return eventId;
     }
 
-    public void setEventId(Long eventId) {
+    public void setEventId(Integer eventId) {
         this.eventId = eventId;
     }
