+++ ServerProject/src/main/java/com/example/vavaplanit/model/repetition/WeeklyRepetition.java
@@ -31,7 +31,7 @@ public class WeeklyRepetition extends Repetition {
     public WeeklyRepetition() {
     }
 
-    public WeeklyRepetition(Long eventId, LocalDate start, LocalDate end, int repetitionInterval,
+    public WeeklyRepetition(Integer eventId, LocalDate start, LocalDate end, int repetitionInterval,
                             Integer daysOfWeek) {
         super(eventId, start, end, repetitionInterval);
         setDaysOfWeek(daysOfWeek);
